Fix warning: '+=': conversion from 'size_t' to 'uint32_t', possible loss of data
Fixed by casting to ``uint32_t``.

+++ Hazel/src/Platform/OpenGL/OpenGLVertexArray.cpp
@@ -68,7 +68,7 @@ namespace Hazel {
 		}
 
 		m_VertexBuffers.push_back(vertexBuffer);
-		m_VertexBufferIndexOffset += layout.GetElements().size();
+		m_VertexBufferIndexOffset += (uint32_t)layout.GetElements().size();
 	}
 
 	void OpenGLVertexArray::SetIndexBuffer(const std::shared_ptr<IndexBuffer>& indexBuffer)
